Add unified Top 10 process analysis functions to query profiler
